requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
azure-core==1.35.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
colorama==0.4.6
cryptography==45.0.6
cymem==2.0.11
databricks-sdk==0.46.0
dbt==1.0.0.38.23
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
feedparser==6.0.12
filelock==3.17.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
gevent==25.9.1
google-auth==2.38.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.3.2
grpcio==1.69.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.4.2
json5==0.10.0
json_repair==0.39.1
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
keras==3.7.0
libclang==18.1.1
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
newsapi-python==0.2.7
nodeenv==1.9.1
numpy==2.0.2
oauthlib==3.2.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-proto==1.31.0
opentelemetry-util-http==0.52b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.3.3
parso==0.8.4
pillow==11.1.0
portalocker==2.10.1
posthog==3.19.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.2
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.3.1
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pywin32==309
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
schema==0.7.7
sgmllib3k==1.0.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tiktoken==0.9.0
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.5
wasabi==1.1.3
wcwidth==0.2.13
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zope.event==6.1
zope.interface==8.2
